+++ regimes/ro/invoices.go
@@ -1,0 +1,41 @@
+package ro
+
+import (
+	"github.com/invopop/gobl/bill"
+	"github.com/invopop/gobl/org"
+	"github.com/invopop/gobl/tax"
+	"github.com/invopop/validation"
+)
+
+func validateInvoice(inv *bill.Invoice) error {
+	simplified := inv.Tags.HasTags(tax.TagSimplified)
+	return validation.ValidateStruct(inv,
+		validation.Field(&inv.Supplier,
+			validation.By(validateInvoiceSupplier),
+			validation.Skip,
+		),
+		validation.Field(&inv.Customer,
+			validation.When(
+				simplified,
+				validation.Nil,
+			).Else(
+				validation.Required,
+			),
+			validation.Skip,
+		),
+	)
+}
+
+func validateInvoiceSupplier(value any) error {
+	p, ok := value.(*org.Party)
+	if !ok || p == nil {
+		return nil
+	}
+	return validation.ValidateStruct(p,
+		validation.Field(&p.TaxID,
+			validation.Required,
+			tax.RequireIdentityCode,
+			validation.Skip,
+		),
+	)
+}